Move vfs classes to plex diretory
