+++ plexfuse/plex/PlexVFSDirEntry.py
