+++ plexfuse/plex/PlexVFSFileEntry.py
